Extract Storage class
